requirements.txt
cycler==0.11.0
fonttools==4.34.4
JPype1==1.4.0
kiwisolver==1.4.3
konlpy==0.6.0
lxml==4.9.1
matplotlib==3.5.2
numpy==1.23.0
packaging==21.3
Pillow==9.2.0
pyparsing==3.0.9
python-dateutil==2.8.2
six==1.16.0
wordcloud==1.8.2.2
